pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "co-ai-redteam"
version = "0.1.0"
description = "AI-powered multi-agent red teaming framework for automated vulnerability discovery, exploitation, and patching"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Co-AI-RedTeam", email = "contact@co-ai-redteam.dev" },
]
keywords = [
    "cybersecurity",
    "red-team",
    "vulnerability-discovery",
    "exploit-generation",
    "security-patching",
    "llm-agents",
    "multi-agent",
    "ai-security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "docker>=7.0.0",
    "numpy>=1.24.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/co-ai-redteam/co-ai-redteam"
Documentation = "https://co-ai-redteam.github.io/co-ai-redteam"
Repository = "https://github.com/co-ai-redteam/co-ai-redteam"
Issues = "https://github.com/co-ai-redteam/co-ai-redteam/issues"
Changelog = "https://github.com/co-ai-redteam/co-ai-redteam/blob/main/CHANGELOG.md"

[project.scripts]
co-redteam = "co_redteam.cli:main"
cart = "co_redteam.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/co_redteam"]

[tool.hatch.build.targets.sdist]
include = ["src/co_redteam", "README.md", "LICENSE", "CHANGELOG.md"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PIE", "T20", "SIM", "RET"]
ignore = ["S101", "S603", "S607", "S108"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106"]
"src/co_redteam/agents/*.py" = ["E501"]  # LLM prompt strings exceed line limits

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=co_redteam --cov-report=term-missing"
